SET_ITEM->NPBonus sutunu için veri tipi düzenlemesi
* Tabloda 8 bitlik veri tipi mevcuttur proje tarafında düzenlendi.

+++ src/GameServer/GameDefine.h
@@ -543,7 +543,7 @@ struct _SET_ITEM
 	uint16 ACBonusClassPercent;	// +AC% for ACBonusClassType only
 
 	uint16 MaxWeightBonus;
-	uint16 NPBonus;
+	uint8 NPBonus;
 };
 
 struct _QUEST_HELPER
